Compare column names in HCES data

This function compares the column names in the hces data file to the standard MAPS column names. It print the changes to the console.

Usage

check_hces_names(data, original_data_df)

Arguments

data

dataframe . The data file as a dataframe.

original_data_df

dataframe . The original data file as a dataframe.

Value

dataframe: The data file with renamed columns, converted data types and a table of the changes to the console.