Get column mapping for a given country and survey

This function returns the column mapping for a given country and survey

get_hces_mappings(country_name, survey_name)

Arguments

country_name

character The country name in ISO3 format.

survey_name

character The survey name in short form.

Value

dataframe: The column mapping for the given country and survey.