Food Unit Names and Unit Codes Supported by this Package

A data frame containing both standard and non-standard food unit names and unit codes of surveys supported by this package.

A data frame containing both standard and non-standard food unit names and unit codes of surveys supported by this package.

unit_names_n_codes_df

unit_names_n_codes_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 208 rows and 6 columns.

A data frame with 3 columns and 4 rows

country

Country name

survey

Survey name

unit_name

Name of the food unit

unit_code

Code of the food unit

\entries_countNumber of survey responses with this unit name \priorityPriority of the unit name (1 being the highest)

Examples

unit_names_n_codes_df
#> # A tibble: 208 × 6
#>    country survey unit_name           unit_code entries_count priority
#>    <chr>   <chr>  <chr>               <chr>             <dbl>    <dbl>
#>  1 MWI     IHS5   100 GRAM            99A                  42        3
#>  2 MWI     IHS5   1000G PACKET        23                    1        4
#>  3 MWI     IHS5   10G PACKET          23                    1        4
#>  4 MWI     IHS5   10KG PAIL           23                    1        4
#>  5 MWI     IHS5   10L BUCKET          81                    4        2
#>  6 MWI     IHS5   125G PACKET         99C                  10        3
#>  7 MWI     IHS5   135G PACKET         23                    1        4
#>  8 MWI     IHS5   149G TIN            23                    2        4
#>  9 MWI     IHS5   1LITRE BOTTLE SUPER 23                    1        4
#> 10 MWI     IHS5   20 LITRE PAIL       4E                    2        2
#> # ℹ 198 more rows